Restructured ChessServer\Socket
