+++ cli/ws.php
@@ -2,7 +2,7 @@
 
 namespace ChessServer\Cli;
 
-use ChessServer\Socket;
+use ChessServer\Socket\Socket;
 use Ratchet\Server\IoServer;
 use Ratchet\Http\HttpServer;
 use Ratchet\WebSocket\WsServer;
